chore: Add PATCH endpoint for updating individual data
This commit adds a new PATCH endpoint in the `tmf632.go` file to handle updating individual data. The endpoint is `/partyManagement/v5/individual/:id` and it calls the `UpdateIndividual` function in the `PartyHandler` struct. This allows users to update individual information by sending a JSON payload with the desired changes. The changes are then converted to SQL update statements using the `JSONconverToUpdateValue` function in the `json_handler.go` file. The SQL statements are executed to update the corresponding fields in the database. This feature enhances the functionality of the party management system by allowing users to modify individual data.